Rename examples files example components to kebab-case
