--- packages/examples/src/components/examples/FilesExample/index.ts
@@ -1,3 +1,0 @@
-export * from './Nav';
-export * from './navItems';
-export * from './Search';
