+++ packages/examples/src/components/examples/files/categories-section.tsx
