+++ packages/examples/src/components/examples/files/favorites-page.tsx
@@ -4,7 +4,7 @@ import {
   forwardRef,
 } from 'react';
 
-import { FileList, type FileListItem } from './FileList';
+import { FileList, type FileListItem } from './file-list';
 
 // Favorites
 const FAVORITE_FILES = [
