+++ packages/examples/src/components/examples/files/index.ts
@@ -1,0 +1,3 @@
+export * from './nav';
+export * from './nav-items';
+export * from './search';
