+++ packages/examples/src/components/examples/files/search.tsx
